requirements/development.txt
django
psycopg2
django-mptt
django-froala-editor
django-debug-toolbar
ipython
pyinotify
flake8
-e git+git@github.com:xtranophilist/django-dbsettings.git@master#egg=django-dbsettings
linaro-django-pagination